+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- DisableAfterAnimationFinish component
 - PlayAndWait animation extension function 
 
 ## [1.1.0] - 2022-10-20